refactor: move realtime related code to directory
