+++ lib/realtime/realtime.js
@@ -12,13 +12,13 @@ const moment = require('moment')
 const get = require('lodash/get')
 
 // core
-const config = require('./config')
-const logger = require('./logger')
-const history = require('./web/history')
-const models = require('./models')
+const config = require('../config')
+const logger = require('../logger')
+const history = require('../web/history')
+const models = require('../models')
 
 // ot
-const ot = require('./ot')
+const ot = require('../ot')
 
 const { ProcessQueue } = require('./processQueue')
 const { RealtimeClientConnection } = require('./realtimeClientConnection')
