+++ lib/realtime/realtimeCleanDanglingUserJob.js
@@ -1,8 +1,8 @@
 'use strict'
 
 const async = require('async')
-const config = require('./config')
-const logger = require('./logger')
+const config = require('../config')
+const logger = require('../logger')
 
 /**
  * clean when user not in any rooms or user not in connected list
