+++ lib/realtime/realtimeClientConnection.js
@@ -2,9 +2,9 @@
 
 const get = require('lodash/get')
 
-const config = require('./config')
-const models = require('./models')
-const logger = require('./logger')
+const config = require('../config')
+const models = require('../models')
+const logger = require('../logger')
 
 class RealtimeClientConnection {
   constructor (socket) {
