+++ lib/realtime/realtimeSaveRevisionJob.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const models = require('./models')
-const logger = require('./logger')
+const models = require('../models')
+const logger = require('../logger')
 
 /**
  * clean when user not in any rooms or user not in connected list
