+++ lib/realtime/realtimeUpdateDirtyNoteJob.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const config = require('./config')
-const logger = require('./logger')
+const config = require('../config')
+const logger = require('../logger')
 const moment = require('moment')
 
 class UpdateDirtyNoteJob {
