+++ lib/web/status/index.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const realtime = require('../../realtime')
+const realtime = require('../../realtime/realtime')
 const config = require('../../config')
 
 exports.getStatus = async (req, res) => {
